% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicModelOD.R
\name{BasicModelOD}
\alias{BasicModelOD}
\title{The Basic Model
The odin.dust SEIRV model, basic version (documentation TBA)}
\description{
The Basic Model
The odin.dust SEIRV model, basic version (documentation TBA)
}
