% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{calc_like0}
\alias{calc_like0}
\title{calc_like0}
\usage{
calc_like0(
  obs_sero_data = NULL,
  obs_case_data = NULL,
  obs_outbreak_data = NULL
)
}
\arguments{
\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{obs_outbreak_data}{Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks,
1 = 1 or more outbreak(s)}
}
\description{
Calculate the likelihood value which would be obtained (with no prior) if modelled data values were
             all equal to observed values (for use as comparison with MCMC likelihood values)
}
\details{
TBA
}
