% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_varFR.R
\name{calc_var_epi}
\alias{calc_var_epi}
\title{calc_var_epi}
\usage{
calc_var_epi(
  coeffs_const = c(),
  coeffs_var = c(),
  enviro_data_const = data.frame(),
  enviro_data_var = list()
)
}
\arguments{
\item{coeffs_const}{TBA}

\item{coeffs_var}{TBA}

\item{enviro_data_const}{TBA}

\item{enviro_data_var}{TBA
'}
}
\description{
Calculate time-varying FOI_spillover or R0 values from environmental covariates and coefficients
}
\details{
TBA
}
