% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_model_varFR.R
\name{case_data_generate_varFR}
\alias{case_data_generate_varFR}
\title{case_data_generate_FR}
\usage{
case_data_generate_varFR(
  FOI_spillover = c(),
  R0 = c(),
  vacc_data = list(),
  pop_data = list(),
  year0 = 1940,
  mode_start = 0,
  n_reps = 1,
  year_end = 2000,
  year_data_begin = 1999,
  vaccine_efficacy = vaccine_efficacy,
  start_SEIRV = list(),
  dt = 1
)
}
\arguments{
\item{FOI_spillover}{= Force of infection due to spillover from sylvatic reservoir (variable by year)}

\item{R0}{= Basic reproduction number for urban spread of infection (variable by year)}

\item{vacc_data}{Vaccination coverage in each age group by year}

\item{pop_data}{Population in each age group by year}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{n_reps}{Number of runs}

\item{year_end}{year to run up to}

\item{year_data_begin}{year to begin saving data}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)
'}
}
\description{
Take in single set of population data and model parameters, output infection data only
}
\details{
Accepts epidemiological + population parameters and model settings; runs full version of SEIRV model
for one region over a specified time period for a number of particles/threads and outputs infection numbers at each
time increment only; optimized for running a large number of repetitions
}
