% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{case_match_graphs}
\alias{case_match_graphs}
\title{case_match_graphs}
\usage{
case_match_graphs(model_data = list(), obs_case_data = list())
}
\arguments{
\item{model_data}{TBA}

\item{obs_case_data}{TBA}
}
\description{
Function to create a series of graphs comparing modelled and observed case data from results
  generated from data_match_multi
}
\details{
TBA
}
