% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{case_match_graphs2}
\alias{case_match_graphs2}
\title{case_match_graphs2}
\usage{
case_match_graphs2(model_data = list(), obs_case_data = list())
}
\arguments{
\item{model_data}{TBA}

\item{obs_sero_data}{TBA}
}
\description{
Function to create a series of graphs comparing modelled and observed case data from results
  generated from data_match2
}
\details{
TBA
}
