% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{cases_compare}
\alias{cases_compare}
\title{cases_compare}
\usage{
cases_compare(model_data = list(), obs_data = list())
}
\arguments{
\item{model_data}{Modelled data in format [TBA]}

\item{obs_data}{Observed annual number of outbreaks}
}
\description{
Compare modelled and observed severe cases using negative binomial
}
\details{
TBA
}
