% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output_process.R
\name{convert_model_output_combine_by_age}
\alias{convert_model_output_combine_by_age}
\title{convert_model_output_combine_by_age}
\usage{
convert_model_output_combine_by_age(model_output = list())
}
\arguments{
\item{model_output}{List of output data produced by appropriate functions
'}
}
\description{
Convert SEIRV output of other functions to combine values across ages
}
\details{
Takes in the output of the Full_Model_Run(), Basic_Model_Run() and similar functions and sums together
S, E, I, R, and V values across age groups, outputting a list containing arrays of values by particle and time point
}
