% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output_process.R
\name{convert_model_output_tidy}
\alias{convert_model_output_tidy}
\title{convert_model_output_tidy}
\usage{
convert_model_output_tidy(model_output = list())
}
\arguments{
\item{model_output}{List of output data produced by appropriate functions
'}
}
\description{
Convert output of Full_Model_Run() or Basic_Model_Run() functions to simple line list
}
\details{
Takes in the output of the Full_Model_Run() or Basic_Model_Run() functions and outputs a data frame
with one set of SEIRV data per line, for use with tidyr functions
}
