% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{create_input_data}
\alias{create_input_data}
\title{create_input_data}
\usage{
create_input_data(vacc_data = list(), pop_data = list())
}
\arguments{
\item{vacc_data}{TBA}

\item{pop_data}{TBA
'}
}
\description{
TBA
}
\details{
TBA
}
