% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create.R
\name{create_map}
\alias{create_map}
\title{create_map}
\usage{
create_map(
  shape_data = list(),
  param_values = c(),
  scale = c(),
  colour_scale = c(),
  pixels_max = 720,
  text_size = 1,
  map_title = "",
  legend_title = "",
  legend_position = "topleft",
  legend_format = "f",
  legend_dp = 1,
  output_file = NULL
)
}
\arguments{
\item{shape_data}{Region shape data generated using map_shapes_load()}

\item{param_values}{Vector of parameter values for regions in shape_data}

\item{scale}{Vector of scale intervals to use for param_values}

\item{colour_scale}{Vector of colours with size greater than or equal to scale - used to convert scale to colours}

\item{pixels_max}{Number of pixels to use for largest dimension of map}

\item{text_size}{Size of text to appear in legend and titles}

\item{map_title}{Title to show above map}

\item{legend_title}{Title to show above legend}

\item{legend_position}{Position to place map legend (select from"bottomright", "bottom", "bottomleft", "left",
"topleft", "top", "topright", "right" and "center")}

\item{legend_format}{Number format to use for scale values in legend - set to "f" (basic number), "e" (scientific
notation) or "pc" (percentage)}

\item{legend_dp}{Number of decimal places to use in scale values in legend (e.g. if legend_format is "f" and
legend_dp is 1, numbers will appear as e.g. 10.0)}

\item{output_file}{Name of file to which to output map; if set to NULL, map is displayed without being saved to file
'}
}
\description{
Create a map of one or more regions with colours denoting parameter values
}
\details{
Takes in region shape data generated using map_shapes_load() and parameter values for each region, plots map
  of the regions and fills regions with colour based on parameter values
}
