% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_param_labels}
\alias{create_param_labels}
\title{create_param_labels}
\usage{
create_param_labels(
  type = "FOI",
  input_data = list(),
  enviro_data = NULL,
  extra_params = c("vacc_eff")
)
}
\arguments{
\item{type}{= Type of MCMC fit to use (should be one of "FOI", "FOI+R0", "FOI enviro" or "FOI+R0 enviro"}

\item{input_data}{= List of population and vaccination data for multiple regions (created using data input creation
code and usually loaded from RDS file)}

\item{enviro_data}{= Environmental data frame, containing only relevant environmental variables}

\item{extra_params}{= Vector of strings listing parameters besides ones determining FOI/R0 (may include vaccine
efficacy and/or infection/death reporting probabilities)}
}
\description{
Apply names to the parameters in the set used for Markov Chain Monte Carlo fitting
}
\details{
TBA
}
