% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{data_match2}
\alias{data_match2}
\title{data_match2}
\usage{
data_match2(
  param_sets = list(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  obs_outbreak_data = NULL,
  const_list = list()
)
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions, with tables to cross-reference
with observed data, added using input_data_process2}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{obs_outbreak_data}{Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks,
1 = 1 or more outbreak(s)}

\item{const_list}{= List of constant parameters/flags/etc. (type,n_reps,mode_start,dt,enviro_data,R0_fixed_values,
vaccine_efficacy,p_obs_severe,p_obs_death)}

\item{param_prop}{Data frame of log values of proposed parameters, one set per row}
}
\description{
Function which runs the model to create simulated data corresponding to supplied observed data for
  multiple parameter sets
}
\details{
TBA
}
