% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{deaths_compare}
\alias{deaths_compare}
\title{deaths_compare}
\usage{
deaths_compare(model_data = list(), obs_data = list())
}
\arguments{
\item{model_data}{Modelled data in data frame format (list of years and number of observed deaths in each)}

\item{obs_data}{Observed annual number of outbreaks
'}
}
\description{
Compare modelled and observed deaths using negative binomial
}
\details{
Compares modelled data on fatal cases per year and compared with observed data, calculating logarithmic
likelihood of observing the latter given the former, using a negative binomial formula.
}
