% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process_multichain.R
\name{display_multichain_progress}
\alias{display_multichain_progress}
\title{display_multichain_progress}
\usage{
display_multichain_progress(
  datasets = list(),
  datasets_selected = c(1),
  burnin_values = NULL,
  end_values = NULL,
  flag_grb = TRUE,
  grb_plot = FALSE
)
}
\arguments{
\item{datasets}{List containing one or more datasets, generated by get_mcmc_datasets_multichain()}

\item{datasets_selected}{Vector of dataset numbers to be selected from list}

\item{burnin_values}{Vector of burnin values to use for selected datasets (set to 1 by default)}

\item{end_values}{Vector of end values to use for selected datasets (set to last value by default)}

\item{flag_grb}{Flag indicating whether to calculate Gelman-Rubin-Brooks convergence metric (MPSRF)}

\item{grb_plot}{Flag indicating whether to plot Gelman-Rubin-Brooks convergence diagnostic (coda::gelman.plot())}
}
\description{
Plot graph of Markov chain progress
}
\details{
This function plots a graph based on one or more datasets generated by get_mcm_datasets_multichain(),
  showing the likelihood as a function of the number of iterations
}
