% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Generate_Dataset}
\alias{Generate_Dataset}
\title{Generate_Dataset}
\usage{
Generate_Dataset(
  input_data = list(),
  enviro_data = list(),
  FOI_values = c(),
  R0_values = c(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  obs_outbreak_data = NULL,
  vaccine_efficacy = 1,
  p_rep_severe = 1,
  p_rep_death = 1,
  mode_start = 1,
  n_reps = 1,
  dt = 1
)
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions (created using data input creation
code and usually loaded from RDS file), with cross-reference tables added using input_data_process() function)}

\item{enviro_data}{TBA}

\item{FOI_values}{TBA}

\item{R0_values}{TBA}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{obs_outbreak_data}{Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks,
1 = 1 or more outbreak(s)}

\item{vaccine_efficacy}{TBA}

\item{p_rep_severe}{TBA}

\item{p_rep_death}{TBA}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{n_reps}{Number of repeats over which to average results}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)
'}
}
\description{
Generate serological, annual case/death and/or annual outbreak risk data based on observed or dummy
  data sets; normally used by single_like_calc() and data_match_single() functions
}
\details{
TBA
}
