% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_generate_split_p_rep.R
\name{Generate_Dataset2}
\alias{Generate_Dataset2}
\title{Generate_Dataset2}
\usage{
Generate_Dataset2(
  input_data = list(),
  FOI_values = c(),
  R0_values = c(),
  sero_template = NULL,
  case_template = NULL,
  vaccine_efficacy = 1,
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe_af = 1,
  p_rep_death_af = 1,
  p_rep_severe_sa = 1,
  p_rep_death_sa = 1,
  mode_start = 1,
  start_SEIRV = NULL,
  dt = 1,
  n_reps = 1,
  deterministic = FALSE,
  mode_parallel = FALSE,
  cluster = NULL,
  output_frame = FALSE
)
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions in standard format; regions should all be
labelled with the 3-letter country code at the beginning so that they can be separated into African and South American ones.
Africa: "AGO","BDI","BEN","BFA","CAF","CIV","CMR","COD","COG","ERI","ETH","GAB","GHA","GIN","GMB","GNB","GNQ",
"KEN","LBR","MLI","MRT","NER","NGA","RWA","SDN","SEN","SLE","SOM","SSD","TCD","TGO","TZA","UGA","ZMB"
South America: "ARG","BOL","BRA","COL","ECU","GUF","GUY","PER","PRY","SUR","VEN"}

\item{FOI_values}{Vector of values of the force of infection due to spillover from sylvatic reservoir}

\item{R0_values}{Vector of values of the basic reproduction number for human-human transmission}

\item{sero_template}{Seroprevalence data template - data frame with region, year, minimum/maximum age, vc_factor [TBA]
and number of samples}

\item{case_template}{Annual reported case/death data template - data frame with region and year}

\item{vaccine_efficacy}{Fractional vaccine efficacy}

\item{p_severe_inf}{Probability of an infection being severe}

\item{p_death_severe_inf}{Probability of a severe infection resulting in death}

\item{p_rep_severe_af}{Probability of reporting of a severe but non-fatal infection (African regions)}

\item{p_rep_death_af}{Probability of a severe infection resulting in death (African regions)}

\item{p_rep_severe_sa}{Probability of reporting of a severe but non-fatal infection (South American regions)}

\item{p_rep_death_sa}{Probability of a severe infection resulting in death (South American regions)}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity (uniform by age, R0 based only)
If mode_start=2, use SEIRV input in list from previous run(s)
If mode_start=3, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age)}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input (list of datasets, one per region)}

\item{dt}{Time increment in days to use in model (should be either 1.0, 2.5 or 5.0 days)}

\item{n_reps}{number of stochastic repetitions}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{mode_parallel}{TRUE/FALSE - set model to run in parallel using cluster if TRUE}

\item{cluster}{Cluster of threads to use if mode_parallel=TRUE}

\item{output_frame}{Flag indicating whether to output a complete data frame of results in template format (if TRUE)
  or calculated values only (if FALSE)
'}
}
\description{
Variant of Generate_Dataset - generate annual serological and/or case/death data with separate case
  reporting probabilities
}
\details{
This function is used to generate annual serological and/or case/death data based on templates;
it is normally used by the single_like_calc() function.

[TBA - Explanation of breakdown of regions to model and how to set lengths of FOI_values and R0_values]
}
