% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{get_mcmc_data}
\alias{get_mcmc_data}
\title{get_mcmc_data}
\usage{
get_mcmc_data(input_folder = "", plot_graph = TRUE, output_data = TRUE)
}
\arguments{
\item{input_folder}{Folder containing CSV files output by mcmc()}

\item{plot_graph}{TRUE/FALSE flag indicating whether to plot results graph}

\item{output_data}{TRUE/FALSE flag indicating whether to output results as data frame}
}
\description{
Read MCMC results from CSV files in folder; plot likelihood graph and/or output data set
}
\details{
TBA
}
