% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{get_MCMC_data_CI}
\alias{get_MCMC_data_CI}
\title{get_MCMC_data_CI}
\usage{
get_MCMC_data_CI(input_frame = list(), margin = 0.95)
}
\arguments{
\item{input_frame}{TBA}

\item{margin}{TBA
'}
}
\description{
Get confidence interval of selected MCMC output parameter values
}
\details{
TBA... Takes logarithms of MCMC output values (actual fitted values) and calculate CI for each parameter,
         then returns exponentials to return to original output format
}
