% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process_multichain.R
\name{get_mcmc_datasets_multichain}
\alias{get_mcmc_datasets_multichain}
\title{get_mcmc_datasets_multichain}
\usage{
get_mcmc_datasets_multichain(input_folders = c())
}
\arguments{
\item{input_folders}{Vector of folders to get chain data}
}
\description{
Get MCMC output data from multiple chains
}
\details{
TBA
}
