% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process_multichain.R
\name{get_mcmc_enviro_coeff_multichain}
\alias{get_mcmc_enviro_coeff_multichain}
\title{get_mcmc_enviro_coeff_multichain}
\usage{
get_mcmc_enviro_coeff_multichain(
  datasets = c(),
  datasets_selected = c(1),
  burnin_values = NULL,
  end_values = NULL,
  type = "FOI+R0",
  enviro_data = NULL
)
}
\arguments{
\item{datasets}{List containing one or more datasets, generated by get_mcmc_datasets_multichain()}

\item{datasets_selected}{Vector of dataset numbers to be selected from list}

\item{burnin_values}{Vector of burn-in iteration values (1 per dataset) before which data discarded}

\item{end_values}{Vector of end iteration values (1 per dataset) before which data discarded}

\item{type}{Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental
covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"}

\item{enviro_data}{Data frame containing values of environmental covariates; set to NULL if not in use}
}
\description{
Get environmental coefficient data from MCMC results
}
\details{
This function extracts values of the coefficients of environmental covariates from datasets
  generated by get_mcmc_datasets_multichain(), returning a list of datasets
}
