% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{get_mcmc_FOI_R0_data}
\alias{get_mcmc_FOI_R0_data}
\title{get_mcmc_FOI_R0_data}
\usage{
get_mcmc_FOI_R0_data(
  input_frame = list(),
  type = "FOI+R0",
  enviro_data = list(),
  plot_type = "none"
)
}
\arguments{
\item{input_frame}{TBA}

\item{type}{TBA}

\item{enviro_data}{TBA - Add note that enviro_data should include ONLY relevant environmental variables and regions}

\item{plot_type}{TBA}
}
\description{
TBA
}
\details{
TBA
}
