% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{get_mcmc_params}
\alias{get_mcmc_params}
\title{get_mcmc_data}
\usage{
get_mcmc_params(chain = list())
}
\arguments{
\item{chain}{Data frame containing MCMC output
'}
}
\description{
Identify columns in MCMC output data containing fitted parameter values and get parameter names
}
\details{
TBA
}
