% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{get_outbreak_data_multi}
\alias{get_outbreak_data_multi}
\title{get_outbreak_data_multi}
\usage{
get_outbreak_data_multi(
  case_data = list(),
  years_data = c(),
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe = 1,
  p_rep_death = 1,
  max_case_interval = 10
)
}
\arguments{
\item{case_data}{Array of cases by time point and repetition summed over age}

\item{years_data}{Vector of year values corresponding to case data}

\item{p_severe_inf}{Probability of an infection being severe}

\item{p_death_severe_inf}{Probability of a severe infection resulting in death}

\item{p_rep_severe}{Probability of a severe infection being reported}

\item{p_rep_death}{Probability of a death being reported}

\item{max_case_interval}{Maximum number of days between reported cases before outbreak is declared over}
}
\description{
Generate outbreak data from model output over multiple repetitions
}
\details{
[TBA]
}
