% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_setup.R
\name{imm_fraction_function_yfd}
\alias{imm_fraction_function_yfd}
\title{imm_fraction_function_yfd}
\usage{
imm_fraction_function_yfd(
  log_lambda = -4,
  R0 = 1,
  ages = c(0:100),
  pop_fraction = rep(1/101, 101)
)
}
\arguments{
\item{log_lambda}{Natural logarithm of force of infection}

\item{R0}{Basic reproduction number}

\item{ages}{List of age values}

\item{pop_fraction}{Population of each age group as proportion of total
'}
}
\description{
Function to estimate notional FOI for herd immunity based on R0 and population age distribution
}
\details{
[TBA]
}
