% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{input_data_check}
\alias{input_data_check}
\title{input_data_check}
\usage{
input_data_check(input_data = list())
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions}
}
\description{
Check that input data is correctly formatted
}
\details{
This function takes in a list of input data for use with other functions and checks that it is correctly
  formatted, including containing all necessary elements and having years and ages in sequence
}
