% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{input_data_truncate}
\alias{input_data_truncate}
\title{input_data_truncate}
\usage{
input_data_truncate(input_data = list(), regions_new = c())
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions (created using create_input_data()
function and usually loaded from an RDS file)}

\item{regions_new}{Vector of regions (subset of input_data$region_labels) for which to create new, shorter dataset}
}
\description{
Truncate input data list for shorter set of regions
}
\details{
TBA
}
