% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create.R
\name{map_shapes_load}
\alias{map_shapes_load}
\title{map_shapes_load}
\usage{
map_shapes_load(regions = c(), shapefiles = c(), region_label_type = "")
}
\arguments{
\item{regions}{TBA}

\item{shapefiles}{TBA}

\item{region_label_type}{TBA
'}
}
\description{
TBA
}
\details{
TBA
}
