% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{MCMC}
\alias{MCMC}
\title{MCMC}
\usage{
MCMC(
  type = NULL,
  pars_ini = c(),
  pars_min = NULL,
  pars_max = NULL,
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  obs_outbreak_data = NULL,
  n_reps = 1,
  Niter = 1,
  mode_start = 0,
  prior_type = "zero",
  dt = 1,
  enviro_data = NULL,
  R0_fixed_values = NULL,
  vaccine_efficacy = NULL,
  p_rep_severe = NULL,
  p_rep_death = NULL,
  filename_prefix = "Chain"
)
}
\arguments{
\item{type}{Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental
covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"}

\item{pars_ini}{Initial log values of parameters}

\item{pars_min}{Lower limits of parameter values if specified}

\item{pars_max}{Upper limits of parameter values if specified}

\item{input_data}{List of population and vaccination data for multiple regions (created using data input creation
code and usually loaded from RDS file)}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no.
cases/no. deaths}

\item{obs_outbreak_data}{Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks,
1 = 1 or more outbreak(s)}

\item{n_reps}{Number of times to repeat calculations to get average likelihood at each step}

\item{Niter}{Total number of steps to run}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start = 0, only vaccinated individuals
If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity}

\item{prior_type}{Text indicating which type of calculation to use for prior probability
If prior_type = "zero", prior probability is always zero
If prior_type = "flat", prior probability is zero if FOI/R0 in designated ranges, -Inf otherwise
If prior_type = "exp", prior probability is given by dexp calculation on FOI/R0 values
If prior_type = "norm", prior probability is given by dnorm calculation on parameter values}

\item{dt}{time increment in days (must be 1 or 5)}

\item{enviro_data}{Values of environmental variables (if in use)}

\item{R0_fixed_values}{Values of R0 to use if not being fitted}

\item{vaccine_efficacy}{Vaccine efficacy (set to NULL if being varied as a parameter)}

\item{p_rep_severe}{Probability of observation of severe infection (set to NULL if being varied as a parameter)}

\item{p_rep_death}{Probability of observation of death (set to NULL if being varied as a parameter)}

\item{filename_prefix}{Prefix of names for output files
'}
}
\description{
Combined MCMC Multi-Region - series of MCMC steps for one or more regions
}
\details{
This is the master function for running a Markov chain to optimize the parameters of the yellow fever model
based on the calculated likelihood of observing supplied data given a particular set of parameters. [TBC]
}
