% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_checks}
\alias{mcmc_checks}
\title{mcmc_checks}
\usage{
mcmc_checks(
  type = NULL,
  pars_ini = c(),
  n_params = 1,
  prior_type = NULL,
  n_regions = 1,
  enviro_data = NULL,
  R0_fixed_values = NULL,
  vaccine_efficacy = NULL,
  p_obs_severe = NULL,
  p_obs_death = NULL
)
}
\arguments{
\item{type}{Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental
covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"}

\item{pars_ini}{= Initial parameter values}

\item{n_params}{= Number of parameters (equal to length of pars_ini)}

\item{prior_type}{= Text indicating which type of calculation to use for prior probability
If prior_type = "zero", prior probability is always zero
If prior_type = "flat", prior probability is zero if FOI/R0 in designated ranges, -Inf otherwise
If prior_type = "exp", prior probability is given by dexp calculation on FOI/R0 values
If prior_type = "norm", prior probability is given by dnorm calculation on parameter values}

\item{n_regions}{= Number of regions}

\item{enviro_data}{= Values of environmental variables (if in use)}

\item{R0_fixed_values}{= Values of R0 to use if not being fitted}

\item{vaccine_efficacy}{= Input value of vaccine efficacy if fixed}

\item{p_obs_severe}{= Input value of severe case reporting probability if fixed}

\item{p_obs_death}{= Input value of fatal case reporting probability if fixed
'}
}
\description{
Perform checks on MCMC inputs
}
\details{
This function, which is called by mcmc(), performs a number of checks on data input to mcmc() to ensure
that that function will run correctly. [TBC]
}
