% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc2.R
\name{MCMC_step2}
\alias{MCMC_step2}
\title{MCMC_step2}
\usage{
MCMC_step2(
  param = c(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  obs_outbreak_data = NULL,
  chain_cov = 1,
  adapt = 0,
  like_current = -Inf,
  const_list = list()
)
}
\arguments{
\item{param}{Log values of parameters}

\item{input_data}{List of population and vaccination data for multiple regions (created using data input creation
code and usually loaded from RDS file)}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{obs_outbreak_data}{Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks,
1 = 1 or more outbreak(s)}

\item{chain_cov}{= Chain covariance}

\item{adapt}{= 0/1 flag indicating which type of calculation to use for proposition value}

\item{like_current}{= Current accepted likelihood value}

\item{const_list}{= List of constant parameters/flags/etc. loaded to mcmc2() (type,pars_min,pars_max,n_reps,
mode_start,prior_type,dt=dt,enviro_data,R0_fixed_values,vaccine_efficacy,p_obs_severe,p_obs_death)}
}
\description{
Single MCMC step - one or more regions
}
\details{
This function runs a single step in a Markov chain set up using the function mcmc(). It proposes a
set of parameters using the param_prop_setup() function, calculates the likelihood of observing the observed data
based on that proposed parameter set, accepts or rejects the proposed parameter set based on the calculated
likelihood and existing chain information, then returns the next line of information for the chain to mcmc(). [TBC]
}
