% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Model_Run_Delay_Many_Reps}
\alias{Model_Run_Delay_Many_Reps}
\title{Model_Run_Delay_Many_Reps}
\usage{
Model_Run_Delay_Many_Reps(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  time_inc = 1,
  n_reps = 1,
  division = 10,
  mode_time = 0
)
}
\arguments{
\item{FOI_spillover}{Vector of values of force of infection due to spillover from sylvatic reservoir
(size depends on mode_time)}

\item{R0}{Vector of values of basic reproduction number for urban spread of infection (size depends on mode_time)}

\item{vacc_data}{Projected vaccination-based immunity (assuming vaccine_efficacy = 1) by age group and year}

\item{pop_data}{Population by age group and year}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination \cr
If mode_start=0, only vaccinated individuals \cr
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age) \cr
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{time_inc}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{n_reps}{Number of repetitions (used to set number of particles and threads)}

\item{division}{Number of particles/threads to run in one go (up to 20)}

\item{mode_time}{Type of time dependence of FOI_spillover and R0 to be used: \cr
 If mode_time = 0, no time variation (constant values)\cr
 If mode_time = 1, FOI/R0 vary annually without seasonality (number of values = number of years to consider) \cr
 If mode_time = 2, FOI/R0 vary with monthly seasonality without inter-annual variation (number of values = 12) \cr
 If mode_time = 3, FOI/R0 vary with daily seasonality without inter-annual variation (number of values = 365/dt) \cr
 If mode_time = 4, FOI/R0 vary annually with monthly seasonality (number of values = 12*number of years to consider) \cr
 If mode_time = 5, FOI/R0 vary annually with daily seasonality (number of values = (365/dt)*number of years to consider)
'}
}
\description{
Run delay SEIRV model for single region for large number of repetitions
}
\details{
Accepts epidemiological + population parameters and model settings; runs delay version of SEIRV model
for one region over a specified time period for a number of repetitions and outputs time-dependent SEIRV
values, infection numbers and/or total force of infection values. Variation of Model_Run_Delay() used for
running a large number of repetitions (>20).

#TODO: Harmonize order of variables between model running, dataset generation, and MCMC functions
}
