% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Model_Run_Reactive}
\alias{Model_Run_Reactive}
\title{Model_Run_Reactive}
\usage{
Model_Run_Reactive(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  vacc_rate_cam = c(),
  t_cam = 0,
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  dt = 1,
  n_particles = 1,
  n_threads = 1,
  deterministic = FALSE,
  response_delay = 56,
  p_rep = c(0, 0),
  outbreak_threshold1 = Inf,
  cluster_threshold1 = Inf
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Basic reproduction number for urban spread of infection}

\item{vacc_data}{[TBA]}

\item{vacc_rate_cam}{[TBA]}

\item{t_cam}{[TBA]}

\item{pop_data}{Population in each age group by year}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{dt}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{n_particles}{number of particles to use}

\item{n_threads}{number of threads to use}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{response_delay}{Delay time in days between a flag being triggered and emergency conditions coming into effect}

\item{p_rep}{Probabilities of an infection being reported as a case under different conditions (TBA)}

\item{outbreak_threshold1}{Threshold total no. reported cases to trigger outbreak flag 1}

\item{cluster_threshold1}{Threshold current infectious fraction to trigger cluster flag 1
'}
}
\description{
Runs reactive version of SEIRV model
}
\details{
Accepts epidemiological + population parameters and model settings; runs reactive SEIRV model
for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV
values, infection numbers and total force of infection values. This version of the model differs from the standard
one in taking two sets of input vaccination data, one the default one and one applied after one or more cases have
been reported. Case reporting is governed by an additional parameter p_rep which can also change after one or more
cases have been reported in order to reflect changes in surveillance.
}
