% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{outbreak_risk_compare}
\alias{outbreak_risk_compare}
\title{outbreak_risk_compare}
\usage{
outbreak_risk_compare(model_outbreak_risk = list(), obs_data = list())
}
\arguments{
\item{model_outbreak_risk}{Modelled annual outbreak risk (0-1)}

\item{obs_data}{Observed annual number of outbreaks}
}
\description{
Calculate likelihood of observed outbreak data based on modelled outbreak risk
}
\details{
TBA
}
