% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{param_calc_enviro}
\alias{param_calc_enviro}
\title{param_calc_enviro}
\usage{
param_calc_enviro(param = c(), enviro_data = c())
}
\arguments{
\item{param}{Parameter values}

\item{enviro_data}{Environmental data frame line, containing only relevant environmental variables
'}
}
\description{
Parameter calculation from environmental variables
}
\details{
Takes in parameter set used for Markov Chain Monte Carlo fitting and calculates values of spillover
force of infection and reproduction number.
}
