% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{param_prop_setup}
\alias{param_prop_setup}
\title{param_prop_setup}
\usage{
param_prop_setup(param = c(), chain_cov = 1, adapt = 0)
}
\arguments{
\item{param}{Previous parameter values used as input}

\item{chain_cov}{Covariance calculated from previous steps in chain}

\item{adapt}{0/1 flag indicating which type of calculation to use for proposition value}
}
\description{
Set up proposed new parameter values for next step in chain
}
\details{
TBA
}
