% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parameter_setup}
\alias{parameter_setup}
\title{Parameter setup}
\usage{
parameter_setup(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  year0 = 1940,
  mode_start = 0,
  year_end = 2000,
  year_data_begin = 1999,
  vaccine_efficacy = 1,
  start_SEIRVC = list(),
  dt = 1
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Reproduction number for urban spread of infection}

\item{vacc_data}{Vaccination coverage in each age group by year}

\item{pop_data}{Population in each age group by year}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRVC input in list from previous run(s)}

\item{year_end}{year to run up to}

\item{year_data_begin}{year to begin saving data}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{start_SEIRVC}{SEIRVC data from end of a previous run to use as input}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)}
}
\description{
Set up parameters to input into model
}
\details{
Takes in multiple inputs, outputs list for use by functions ()
}
