% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{plot_mcmc_enviro_coeff_data}
\alias{plot_mcmc_enviro_coeff_data}
\title{plot_mcmc_enviro_coeff_data}
\usage{
plot_mcmc_enviro_coeff_data(
  data_frame = list(),
  env_vars = c(),
  plot_type = "box",
  text_size1 = 10
)
}
\arguments{
\item{data_frame}{Data frame of coefficient values obtained using get_mcmc_enviro_coeff_data()}

\item{env_vars}{List of environmental covariates}

\item{plot_type}{Type of plots to create (choose from "box","violin")}

\item{text_size1}{Text size parameter for axis labels
'}
}
\description{
Plot environmental coefficients from MCMC results
}
\details{
This function takes in a data frame of values (obtained using the get_mcmc_enviro_coeff_data() function) of
  the coefficients of environmental covariates and plots them on graphs
}
