% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{plot_mcmc_FOI_R0_data}
\alias{plot_mcmc_FOI_R0_data}
\title{plot_mcmc_FOI_R0_data}
\usage{
plot_mcmc_FOI_R0_data(
  data_frame,
  regions,
  plot_type = "box",
  output_file_prefix = NULL,
  margin = 0.95
)
}
\arguments{
\item{data_frame}{TBA}

\item{regions}{TBA}

\item{plot_type}{TBA}

\item{output_file_prefix}{TBA}

\item{margin}{TBA}
}
\description{
TBA
}
\details{
TBA
}
