% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{plot_mcmc_prob_data}
\alias{plot_mcmc_prob_data}
\title{plot_mcmc_prob_data}
\usage{
plot_mcmc_prob_data(
  input_frame,
  plot_type = "box",
  values = c("vacc_eff"),
  margin = 0.95
)
}
\arguments{
\item{input_frame}{TBA}

\item{plot_type}{TBA}

\item{values}{TBA}

\item{margin}{TBA}
}
\description{
TBA
}
\details{
TBA
}
