% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_model_output}
\alias{plot_model_output}
\title{plot_model_output}
\usage{
plot_model_output(model_output = list())
}
\arguments{
\item{model_output}{List of output data produced via Full_Model_Run() or Basic_Model_Run() functions
'}
}
\description{
Plot output of Full_Model_Run() or Basic_Model_Run() functions
}
\details{
Takes in the output of the Full_Model_Run() or Basic_Model_Run() functions and outputs a ggplot graph of
  S, R and V summed over all age groups, with error bars showing the spread of values from multiple repetitions
}
