% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_model.R
\name{Reactive_Model_Run}
\alias{Reactive_Model_Run}
\title{Reactive_Model_Run}
\usage{
Reactive_Model_Run(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data1 = list(),
  vacc_data2 = list(),
  pop_data = list(),
  year0 = 1940,
  mode_start = 0,
  n_particles = 1,
  n_threads = 1,
  year_end = 2000,
  year_data_begin = 1999,
  vaccine_efficacy = 1,
  start_SEIRV = list(),
  dt = 1,
  p_rep = c(1e-06, 1e-06),
  outbreak_threshold1 = 1,
  cluster_threshold1 = 1
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Reproduction number for urban spread of infection}

\item{vacc_data1}{Vaccination coverage in each age group by year (non-emergency)}

\item{vacc_data2}{Vaccination coverage in each age group by year (emergency)}

\item{pop_data}{Population in each age group by year}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRVC input in list from previous run(s)}

\item{n_particles}{number of particles to use}

\item{n_threads}{number of threads to use}

\item{year_end}{year to run up to}

\item{year_data_begin}{year to begin saving data}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)}

\item{p_rep}{Probabilities of an infection being reported as a case under different conditions (TBA)}

\item{outbreak_threshold1}{Threshold total no. reported cases to trigger outbreak flag 1}

\item{cluster_threshold1}{Threshold current infectious fraction to trigger cluster flag 1}
}
\description{
Run reactive version of SEIRV model
}
\details{
Accepts epidemiological + population parameters and model settings; runs full version of SEIRV model
for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV
values, infection numbers and total force of infection values. Alternate version incorporating case reporting
and reactive surveillance/control measures based on case numbers
}
