% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero.R
\name{sero_calculate}
\alias{sero_calculate}
\title{sero_calculate}
\usage{
sero_calculate(
  age_min = 0,
  age_max = 101,
  years = NULL,
  vc_factor = 0,
  data = list()
)
}
\arguments{
\item{age_min}{= Minimum age of age group}

\item{age_max}{= Maximum age of age group}

\item{years}{= Years for which to calculate average annual seroprevalence}

\item{vc_factor}{= Proportion of patients tested for whom vaccine status unknown}

\item{data}{= Output of Basic_Model_Run or Full_Model_Run}
}
\description{
Calculate seroprevalence in unvaccinated people from modelled data for one or more years and one age
range
}
\details{
TBA
}
