% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero.R
\name{sero_compare}
\alias{sero_compare}
\title{sero_compare}
\usage{
sero_compare(model_data = list(), obs_sero_data = list(), n_particle = 1)
}
\arguments{
\item{model_data}{= Output of Basic_Model_Run_OD or Full_Model_Run_OD}

\item{obs_sero_data}{= Seroprevalence data for comparison, by year and age group, in format
no. samples/no. positives}

\item{n_particle}{= Particle to select from model_data
'}
}
\description{
Take model results, calculate seroprevalence for comparison with observed seroprevalence and
calculate likelihood (single region, multiple years/age ranges)
}
\details{
Takes in SEIRV model output data and observed seroprevalence data, calculates seroprevalence from modelled
data, and compares modelled and observed data, calculating logarithmic likelihood of observing the latter given the
former, using a binomial formula.
}
