% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero.R
\name{sero_compare_multiparticle}
\alias{sero_compare_multiparticle}
\title{sero_compare_multiparticle}
\usage{
sero_compare_multiparticle(model_data = list(), obs_sero_data = list())
}
\arguments{
\item{model_data}{= Output of Basic_Model_Run or Full_Model_Run}

\item{obs_sero_data}{= Seroprevalence data for comparison, by year and age group, in format no. samples/no.
  positives
'}
}
\description{
Take model results from multi-particle run using OD version of model, calculate
seroprevalence for comparison with observed seroprevalence  and  calculate likelihood value (single region,
multiple years/age ranges) for each particle; return as vector of values
}
\details{
Calculates logarithmic likelihood of observing observed seroprevalence given modelled data, similar to
sero_compare(), but using modelled data for multiple particles.
}
