% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{sero_match_graphs}
\alias{sero_match_graphs}
\title{sero_match_graphs}
\usage{
sero_match_graphs(model_data = list(), obs_sero_data = list(), type = "mean")
}
\arguments{
\item{model_data}{TBA}

\item{obs_sero_data}{TBA}

\item{type}{TBA}
}
\description{
Function to create a series of graphs comparing modelled and observed serological data from results
  generated from data_match_multi
}
\details{
TBA
}
