% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{sero_match_graphs2}
\alias{sero_match_graphs2}
\title{sero_match_graphs2}
\usage{
sero_match_graphs2(model_data = list(), obs_sero_data = list())
}
\arguments{
\item{model_data}{TBA}

\item{obs_sero_data}{TBA}
}
\description{
Function to create a series of graphs comparing modelled and observed serological data from results
  generated from data_match2
}
\details{
TBA
}
