% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc2.R
\name{single_like_calc2}
\alias{single_like_calc2}
\title{single_like_calc2}
\usage{
single_like_calc2(
  param_prop = c(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  obs_outbreak_data = NULL,
  const_list = list()
)
}
\arguments{
\item{param_prop}{Log values of proposed parameters}

\item{input_data}{List of population and vaccination data for multiple regions (created using data input creation
code and usually loaded from RDS file), with cross-reference tables added using input_data_process2 in MCMC2}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{obs_outbreak_data}{Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks,
1 = 1 or more outbreak(s)}

\item{const_list}{= List of constant parameters/flags/etc. loaded to mcmc2() (type,pars_min,pars_max,n_reps,
mode_start,prior_type,dt=dt,enviro_data,R0_fixed_values,vaccine_efficacy,p_obs_severe,p_obs_death)}
}
\description{
Function which calculates and outputs likelihood
}
\details{
This function calculates the total likelihood of observing a set of observations (across multiple regions
and data types) for a given proposed parameter set. [TBC]
}
