% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_varFR.R
\name{single_posterior_calc_VarFR}
\alias{single_posterior_calc_VarFR}
\title{single_posterior_calc_VarFR}
\usage{
single_posterior_calc_VarFR(
  log_params_prop = c(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  ...
)
}
\arguments{
\item{log_params_prop}{Proposed values of parameters to be estimated (natural logarithm of actual parameters)}

\item{input_data}{List of population and vaccination data for multiple regions (created using data input
creation code and usually loaded from RDS file), with cross-reference tables added using input_data_process
in MCMC}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{...}{= Constant parameters/flags/etc. loaded to or determined by mcmc() and mcmc_prelim_fit, including mode_start,
prior_settings, dt, n_reps, enviro_data, p_severe_inf, p_death_severe_inf, add_values list, extra_estimated_params,
deterministic, mode_time, mode_parallel, cluster, i_FOI_const, i_FOI_var, i_R0_const, i_R0_var}
}
\description{
Function which calculates and outputs posterior likelihood of observing simulated data
}
\details{
This function calculates the posterior likelihood of observing a set of observations (across multiple
regions and data types) for a given proposed parameter set. [TBA]
}
