% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{total_burden_estimate}
\alias{total_burden_estimate}
\title{total_burden_estimate}
\usage{
total_burden_estimate(
  type = "FOI+R0 enviro",
  param_dist = list(),
  input_data = list(),
  start_SEIRV0 = NULL,
  years_data = c(),
  n_reps = 1,
  mode_start = 1,
  dt = 5,
  enviro_data = NULL,
  R0_fixed_values = NULL,
  vaccine_efficacy0 = NULL,
  p_rep_severe0 = NULL,
  p_rep_death0 = NULL,
  flag_reporting = TRUE
)
}
\arguments{
\item{type}{TBA}

\item{param_dist}{TBA}

\item{input_data}{TBA}

\item{start_SEIRV0}{TBA}

\item{years_data}{TBA}

\item{n_reps}{TBA}

\item{mode_start}{TBA}

\item{dt}{TBA}

\item{enviro_data}{TBA}

\item{R0_fixed_values}{TBA}

\item{vaccine_efficacy0}{TBA}

\item{p_rep_severe0}{TBA}

\item{p_rep_death0}{TBA}

\item{flag_reporting}{TRUE/FALSE flag indicating whether to calculate number of reported/observed cases}
}
\description{
TBA
}
\details{
TBA
}
