% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_model_varFR.R
\name{total_burden_estimate_varFR}
\alias{total_burden_estimate_varFR}
\title{total_burden_estimate_varFR}
\usage{
total_burden_estimate_varFR(
  FOI_spillover_values = list(),
  R0_values = list(),
  input_data = list(),
  start_SEIRV = NULL,
  years_data = c(),
  n_reps = 1,
  mode_start = 1,
  flag_reporting = FALSE,
  dt = 5,
  vaccine_efficacy = NULL,
  p_rep_severe = NULL,
  p_rep_death = NULL
)
}
\arguments{
\item{FOI_spillover_values}{#Array of values of spillover force of infection by region (dim1) and year (dim2)}

\item{R0_values}{# #Array of values of R0 by region (dim1) and year (dim2)}

\item{input_data}{List of population and vaccination data for multiple regions}

\item{start_SEIRV}{SEIRV data to use as input}

\item{years_data}{Vector of years for which to output data}

\item{n_reps}{Number of repeats over which to average results}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{flag_reporting}{Flag indicating whether to output number of reported severe and fatal cases}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)}

\item{vaccine_efficacy}{Vaccine efficacy (set to NULL if being varied as a parameter)}

\item{p_rep_severe}{Probability of observation of severe infection (set to NULL if being varied as a parameter)}

\item{p_rep_death}{Probability of observation of death (set to NULL if being varied as a parameter)}
}
\description{
Function to calculate annual yellow fever burden across multiple regions based on derived parameters
  (variation for annually varying FOI_spillover and R0)
}
\details{
Function to take in parameter sets derived from MCMC fitting and use to calculate annual total and reported
case and death numbers for multiple regions to compare with external data
}
