% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{truncate_mcmc_data}
\alias{truncate_mcmc_data}
\title{truncate_mcmc_data}
\usage{
truncate_mcmc_data(
  input_frame = list(),
  selection = c(1),
  plot_graph = TRUE,
  output_data = TRUE
)
}
\arguments{
\item{input_frame}{Data frame of output data}

\item{selection}{Vector of line numbers to extract from input_frame}

\item{plot_graph}{TRUE/FALSE flag indicating whether to plot results graph}

\item{output_data}{TRUE/FALSE flag indicating whether to output results as data frame}
}
\description{
Take MCMC results generated by get_mcmc_data and extract selected lines; optionally plot likelihood
graph and/or output data set
}
\details{
TBA
}
