% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{cases_compare}
\alias{cases_compare}
\title{cases_compare}
\usage{
cases_compare(model_data = list(), obs_data = list())
}
\arguments{
\item{model_data}{Modelled data in data frame format (list of years and number of reported severe cases in each)}

\item{obs_data}{Data frame containing annual observed case data
'}
}
\description{
Compare modelled and observed severe cases using negative binomial
}
\details{
Compares modelled data on severe cases per year and compared with observed data, calculating logarithmic
likelihood of observing the latter given the former, using a negative binomial formula.
}
