% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{deaths_compare}
\alias{deaths_compare}
\title{deaths_compare}
\usage{
deaths_compare(model_data = list(), obs_data = list())
}
\arguments{
\item{model_data}{Modelled data in format [TBA]}

\item{obs_data}{Observed annual number of outbreaks}
}
\description{
Compare modelled and observed deaths using negative binomial
}
\details{
TBA
}
