% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process_multichain.R
\name{get_mcmc_probs_multichain}
\alias{get_mcmc_probs_multichain}
\title{get_mcmc_probs_multichain}
\usage{
get_mcmc_probs_multichain(
  datasets = c(),
  datasets_selected = c(1),
  burnin_values = NULL,
  end_values = NULL,
  prob_names = c("vaccine_efficacy")
)
}
\arguments{
\item{datasets}{List containing one or more datasets, generated by get_mcmc_datasets_multichain()}

\item{datasets_selected}{Vector of dataset numbers to be selected from list}

\item{burnin_values}{TBA}

\item{end_values}{TBA}

\item{prob_names}{TBA}
}
\description{
TBA
}
\details{
TBA
}
