% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{get_outbreak_data}
\alias{get_outbreak_data}
\title{get_outbreak_data}
\usage{
get_outbreak_data(
  case_data = c(),
  year_data = c(),
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe = 1,
  p_rep_death = 1
)
}
\arguments{
\item{case_data}{Vector of cases by time point summed over age}

\item{year_data}{Vector of year values corresponding to case data}

\item{p_severe_inf}{Probability of an infection being severe}

\item{p_death_severe_inf}{Probability of a severe infection resulting in death}

\item{p_rep_severe}{probability of a severe infection being reported}

\item{p_rep_death}{probability of a death being reported
'}
}
\description{
Generate outbreak data from model output
}
\details{
Takes in case (infection) data output by Full_Model_OD() or case_data_generate(), uses severe and fatal
infection rates and reporting probabilities to calculate the number of reported cases, and outputs a list of
reported outbreaks. An outbreak is assumed to be reported when one or more cases is reported, and continues until
no new cases have been reported for 10 days.
}
