% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{input_data_process}
\alias{input_data_process}
\title{input_data_process}
\usage{
input_data_process(
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  obs_outbreak_data = NULL
)
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions (created using create_input_data()
function and usually loaded from an RDS file)}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{obs_outbreak_data}{Outbreak Y/N data for comparison, by region and year, in format 0 = no outbreaks,
1 = 1 or more outbreak(s)}
}
\description{
Cross-reference input data with serological, annual case/death and/or outbreak data for comparison
}
\details{
This function, used to prepare input data for functions used to calculate the likelihood of observed data,
  amends a list of population and vaccination data used as input for other functions,
  cross-referencing it with seroprevalence, case and/or outbreak data and adding connection information.
}
