% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_checks}
\alias{mcmc_checks}
\title{mcmc_checks}
\usage{
mcmc_checks(
  log_params_ini = c(),
  n_regions = 1,
  type = NULL,
  log_params_min = c(),
  log_params_max = c(),
  prior_type = NULL,
  enviro_data = NULL,
  R0_fixed_values = NULL,
  vaccine_efficacy = NULL,
  p_rep_severe = NULL,
  p_rep_death = NULL,
  m_FOI_Brazil = 1
)
}
\arguments{
\item{log_params_ini}{Initial values of varied parameters (natural logarithm of actual parameters)}

\item{n_regions}{Number of regions}

\item{type}{Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental
covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"}

\item{log_params_min}{Lower limits of varied parameter values if specified}

\item{log_params_max}{Upper limits of varied parameter values if specified}

\item{prior_type}{Text indicating which type of calculation to use for prior probability
If prior_type = "zero", prior probability is always zero
If prior_type = "flat", prior probability is zero if FOI/R0 in designated ranges, -Inf otherwise
If prior_type = "exp", prior probability is given by dexp calculation on FOI/R0 values
If prior_type = "norm", prior probability is given by dnorm calculation on parameter values}

\item{enviro_data}{Values of environmental covariates (if in use)}

\item{R0_fixed_values}{Values of R0 to use if not being varied (set to NULL if R0 is being varied)}

\item{vaccine_efficacy}{Vaccine efficacy (set to NULL if being varied as a parameter)}

\item{p_rep_severe}{Probability of observation of severe infection (set to NULL if being varied as a parameter)}

\item{p_rep_death}{Probability of observation of death (set to NULL if being varied as a parameter)}

\item{m_FOI_Brazil}{FOI multiplier for Brazil regions}
}
\description{
Perform checks on MCMC inputs
}
\details{
This function, which is called by MCMC(), performs a number of checks on data to be used in fitting to
ensure proper functionality. It verifies that the number of parameters being varied is consistent with other
settings and that certain values are not outwith sensible boundaries (e.g. probabilities must be between 0 and 1).
}
