% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_FOI_R0_setup}
\alias{mcmc_FOI_R0_setup}
\title{mcmc_FOI_R0_setup}
\usage{
mcmc_FOI_R0_setup(
  type = "",
  prior_type = "",
  regions = "",
  log_params_prop = c(),
  enviro_data = list(),
  R0_fixed_values = c(),
  log_params_min = c(),
  log_params_max = c()
)
}
\arguments{
\item{type}{Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental
covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"}

\item{prior_type}{Text indicating which type of calculation to use for prior probability
If prior_type = "zero", prior probability is always zero
If prior_type = "flat", prior probability is zero if FOI/R0 in designated ranges, -Inf otherwise
If prior_type = "exp", prior probability is given by dexp calculation on FOI/R0 values
If prior_type = "norm", prior probability is given by dnorm calculation on parameter values}

\item{regions}{Vector of region names}

\item{log_params_prop}{Proposed values of varied parameters (natural logarithm of actual parameters)}

\item{enviro_data}{Environmental data frame, containing only relevant environmental variables}

\item{R0_fixed_values}{Values of R0 to use if not being varied}

\item{log_params_min}{Lower limits of varied parameter values if specified}

\item{log_params_max}{Upper limits of varied parameter values if specified
'}
}
\description{
Set up FOI and R0 values and calculate some prior probability values for MCMC calculation
}
\details{
Takes in parameter values used for Markov Chain Monte Carlo fitting, calculates spillover force of
infection and (optionally) reproduction number values either directly or from environmental covariates. Also
calculates related components of prior probability.
}
