% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Model_Run_Reactive}
\alias{Model_Run_Reactive}
\title{Model_Run_Delay}
\usage{
Model_Run_Reactive(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  dt = 1,
  n_particles = 1,
  n_threads = 1,
  deterministic = FALSE,
  p_rep = 1,
  outbreak_threshold1 = 1,
  cluster_threshold1 = 1
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Basic reproduction number for urban spread of infection}

\item{vacc_data}{Vaccination coverage in each age group by year}

\item{pop_data}{Population in each age group by year}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{dt}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{n_particles}{number of particles to use}

\item{n_threads}{number of threads to use}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{p_rep}{Probabilities of an infection being reported as a case under different conditions (TBA)}

\item{outbreak_threshold1}{Threshold total no. reported cases to trigger outbreak flag 1}

\item{cluster_threshold1}{Threshold current infectious fraction to trigger cluster flag 1
'}
}
\description{
Runs reactive version of SEIRV model
}
\details{
Accepts epidemiological + population parameters and model settings; runs SEIRV model
for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV
values, infection numbers and total force of infection values. Alternate version incorporating case reporting
and reactive surveillance/control measures based on case numbers
}
