% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Model_Run_Reactive}
\alias{Model_Run_Reactive}
\title{Model_Run_Reactive}
\usage{
Model_Run_Reactive(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  time_inc = 1,
  n_particles = 1,
  n_threads = 1,
  deterministic = FALSE,
  response_delay = 56,
  p_rep = c(0, 0),
  case_threshold = Inf,
  cluster_threshold = Inf,
  vacc_cov_cam = c(),
  t_cam = 0
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Basic reproduction number for urban spread of infection}

\item{vacc_data}{Projected vaccination-based immunity (assuming vaccine_efficacy=1 and in absence of emergency campaign)
by age group and year}

\item{pop_data}{Population by age group and year}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{start_SEIRV}{SEIRV data (including E_delay and I_delay) from end of a previous run to use as input}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{time_inc}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{n_particles}{number of particles to use}

\item{n_threads}{number of threads to use}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{response_delay}{Delay time in days between a threshold being reached and emergency conditions coming into effect}

\item{p_rep}{Probabilities of an infection being reported as a case before emergency conditions triggered (1st value) or
after emergency conditions triggered (2nd value)}

\item{case_threshold}{Threshold total no. reported cases to trigger emergency conditions}

\item{cluster_threshold}{Threshold current infectious fraction to trigger emergency conditions}

\item{vacc_cov_cam}{Target vaccination coverage by age group during emergency vaccination campaign}

\item{t_cam}{Duration in days of emergency vaccination campaign
'}
}
\description{
Runs reactive version of SEIRV model
}
\details{
Accepts epidemiological + population parameters and model settings; runs reactive SEIRV model
for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV
values, infection numbers and total force of infection values. This version of the model differs from the standard
one in simulating an emergency vaccination campaign applied when an outbreak is declared. Case reporting is
governed by an additional parameter p_rep which can also change after a reported outbreak is triggered in order
to reflect changes in surveillance. An outbreak is declared when the number of reported cases or the infected
fraction of the population exceed supplied thresholds.
}
