% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Model_Run_Split}
\alias{Model_Run_Split}
\title{Model_Run_Split}
\usage{
Model_Run_Split(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  dt = 1,
  n_particles = 1,
  n_threads = 1,
  deterministic = FALSE
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Basic reproduction number for urban spread of infection}

\item{vacc_data}{Projected vaccination-based immunity (assuming vaccine_efficacy=1) by age group and year}

\item{pop_data}{Population by age group and year}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{dt}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{n_particles}{number of particles to use}

\item{n_threads}{number of threads to use}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE
'}
}
\description{
Run full SEIRV model with daily infection output split into sylvatic and urban
}
\details{
Accepts epidemiological + population parameters and model settings; runs split-infection SEIRV model
for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV
values, infection numbers and total force of infection values. This version of the model differs from the standard
one in that infections are split into sylvatic and urban, allowing the relative importance of sylvatic and urban
infections to be assessed.
}
