% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Model_Run_VarFR}
\alias{Model_Run_VarFR}
\title{Model_Run_VarFR}
\usage{
Model_Run_VarFR(
  FOI_spillover = c(),
  R0 = c(),
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  output_type = "full",
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  time_inc = 1,
  n_particles = 1,
  n_threads = 1,
  deterministic = FALSE,
  mode_time = 1
)
}
\arguments{
\item{FOI_spillover}{Vector of values of force of infection due to spillover from sylvatic reservoir (size depends on mode_time)}

\item{R0}{Vector of values of basic reproduction number for urban spread of infection (size depends on mode_time)}

\item{vacc_data}{Projected vaccination-based immunity (assuming vaccine_efficacy=1) by age group and year}

\item{pop_data}{Population by age group and year}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{output_type}{Type of data to output:
"full" = SEIRVC + FOI for all steps and ages
"case" = annual total new infections (C) summed across all ages
"sero" = annual SEIRV
"case+sero" = annual SEIRVC, cases summed across all ages
"case_alt" = annual total new infections not combined by age
"case_alt2" = total new infections combined by age for all steps}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start = 0, only vaccinated individuals
If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity (uniform by age,
R0 based only)
If mode_start = 2, use SEIRV input in list from previous run(s)
If mode_start = 3, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{time_inc}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{n_particles}{number of particles to use}

\item{n_threads}{number of threads to use}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{mode_time}{Type of time dependence of FOI_spillover and R0 to be used:
 If mode_time=0, no time variation (identical to Model_Run())
 If mode_time=1, FOI/R0 vary annually without seasonality (number of values = number of years to consider)
 If mode_time=2, FOI/R0 vary with monthly seasonality without inter-annual variation (number of values = 12)
 If mode_time=3, FOI/R0 vary with daily seasonality without inter-annual variation (number of values = 365/time_inc)
 If mode_time=4, FOI/R0 vary annually with monthly seasonality (number of values = 12*number of years to consider)
 If mode_time=5, FOI/R0 vary annually with daily seasonality (number of values = (365/time_inc)*number of years to consider)
'}
}
\description{
Run alternate version of SEIRV model with annually varying FOI_spillover and R0
}
\details{
Accepts epidemiological + population parameters and model settings; runs version of SEIRV model with
time-varying FOI_spillover and R0 values for one region over a specified time period for a number of
particles/threads and outputs time-dependent SEIRV values, infection numbers and total force of infection values.
}
