% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{param_prop_setup}
\alias{param_prop_setup}
\title{param_prop_setup}
\usage{
param_prop_setup(log_params = c(), chain_cov = 1, adapt = 0)
}
\arguments{
\item{log_params}{Previous log parameter values used as input}

\item{chain_cov}{Covariance calculated from previous steps in chain}

\item{adapt}{0/1 flag indicating which type of calculation to use for proposition value
'}
}
\description{
Set up proposed new log parameter values for next step in chain
}
\details{
Takes in current values of parameter set used for Markov Chain Monte Carlo fitting and proposes new values
from multivariate normal distribution where the existing values form the mean and the standard deviation is
based on the chain covariance or (if the flag "adapt" is set to 1) a flat value based on the number of parameters.
}
