% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{plot_mcmc_prob_data}
\alias{plot_mcmc_prob_data}
\title{plot_mcmc_prob_data}
\usage{
plot_mcmc_prob_data(
  input_frame = list(),
  plot_type = "box",
  values = c("vaccine_efficacy"),
  text_size1 = 10
)
}
\arguments{
\item{input_frame}{Data frame of MCMC output data}

\item{plot_type}{Type of plots to create (choose from "box","violin","error_bars")}

\item{values}{List of names of parameters to plot (must be parameters appearing in the data)}

\item{text_size1}{Text size parameter for axis labels}
}
\description{
Plot values of vaccine efficacy and/or reporting probabilities from MCMC output data
}
\details{
This function takes in a data frame produced by functions get_mcmc_data() and/or truncate_mcmc_data(),
calculates the values of additional parameters fitted using the function besides spillover force of infection (FOI)
and reproduction number (R0) plots the values in one of a number of different ways to display the spread of values
obtained (box plots, violin plots, simple plots with error bars).
}
