% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output_process.R
\name{plot_model_output}
\alias{plot_model_output}
\title{plot_model_output}
\usage{
plot_model_output(model_output = list())
}
\arguments{
\item{model_output}{List of SEIRV output data produced by appropriate functions
'}
}
\description{
Plot outputs of SEIRV model combined across age groups
}
\details{
Takes in SEIRV output from Full_Model_Run(), Basic_Model_Run() and similar functions and outputs a ggplot
  graph of S, R and V summed over all age groups, with error bars showing the spread of values from multiple
  particles. If data for a single particle is selected, E and I values are also shown.
}
