% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero.R
\name{sero_calculate}
\alias{sero_calculate}
\title{sero_calculate}
\usage{
sero_calculate(
  age_min = 0,
  age_max = 101,
  years = NULL,
  vc_factor = 0,
  data = list(),
  n_p = 1
)
}
\arguments{
\item{age_min}{= Minimum age of age group}

\item{age_max}{= Maximum age of age group}

\item{years}{= Years for which to calculate average annual seroprevalence}

\item{vc_factor}{= Proportion of patients tested for whom vaccine status unknown}

\item{data}{= Output of Basic_Model_Run or Full_Model_Run}

\item{n_p}{= Particle to select from data
'}
}
\description{
Calculate seroprevalence in unvaccinated people from modelled data for one or more years and one age
range
}
\details{
Takes in information on minimum and maximum ages of desired range, year(s) for which to calculate
seroprevalence, factor representing proportion of patients with unknown vaccine status, and SEIRV model output
data, and calculates seroprevalence in unvaccinated people in specified age range for specified year(s).
}
