% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero.R
\name{sero_compare}
\alias{sero_compare}
\title{sero_compare}
\usage{
sero_compare(model_data = list(), obs_sero_data = list())
}
\arguments{
\item{model_data}{= Output of Basic_Model_Run or Full_Model_Run}

\item{obs_sero_data}{= Seroprevalence data for comparison, by year and age group, in format
no. samples/no. positives}
}
\description{
Take model results, calculate seroprevalence for comparison with observed seroprevalence and
calculate likelihood (single region, multiple years/age ranges)
}
\details{
TBA
}
