% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{sero_match_graphs}
\alias{sero_match_graphs}
\title{sero_match_graphs}
\usage{
sero_match_graphs(
  model_data = list(),
  obs_sero_data = list(),
  plot_type = "mean",
  text_size = 1,
  hide_observed = FALSE
)
}
\arguments{
\item{model_data}{Simulated datasets produced by data_match_multi() function}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{plot_type}{Form in which to plot model data: "all": bands showing 95\% and 50\% of all values;
"mean": bands showing 95\% and 50\% confidence intervals for mean of all values}

\item{text_size}{Size of text to display on graphs}

\item{hide_observed}{If TRUE, indicates that supplied observed data is "dummy" data only supplied to indicate years
and age ranges, and should not be plotted on the graph(s)}
}
\description{
Function to create a series of graphs comparing modelled and observed serological data from results
  generated from data_match_multi() function
}
\details{
Takes in simulated datasets produced by data_match_multi() function along with observed seroprevalence
  dataset used to generate the simulated datasets, and produces graph(s) (one graph per survey) showing the observed
  and simulated data on the same set(s) of axes, with observed data displayed as points and lines and simulated data
  as coloured bands.
}
