% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{total_burden_estimate}
\alias{total_burden_estimate}
\title{total_burden_estimate}
\usage{
total_burden_estimate(
  type = "FOI+R0 enviro",
  param_dist = list(),
  input_data = list(),
  start_SEIRV0 = NULL,
  years_data = c(),
  n_reps = 1,
  mode_start = 1,
  dt = 5,
  enviro_data = NULL,
  R0_fixed_values = NULL,
  vaccine_efficacy0 = NULL,
  p_rep_severe0 = NULL,
  p_rep_death0 = NULL,
  flag_reporting = TRUE
)
}
\arguments{
\item{type}{Type of parameter set (FOI only, FOI+R0, FOI and/or R0 coefficients associated with environmental
covariates); choose from "FOI","FOI+R0","FOI enviro","FOI+R0 enviro"}

\item{param_dist}{Data frame of log values of input parameters, one set per row}

\item{input_data}{List of population and vaccination data for multiple regions}

\item{start_SEIRV0}{SEIRV data to use as input}

\item{years_data}{Vector of years for which to output data}

\item{n_reps}{Number of repeats over which to average results}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)}

\item{enviro_data}{enviro_data Data frame containing values of environmental covariates; set to NULL if not in use}

\item{R0_fixed_values}{Values of R0 to use if not being taken from parameter distribution}

\item{vaccine_efficacy0}{Vaccine efficacy (set to NULL if being varied as a parameter)}

\item{p_rep_severe0}{Probability of observation of severe infection (set to NULL if being varied as a parameter)}

\item{p_rep_death0}{Probability of observation of death (set to NULL if being varied as a parameter)}

\item{flag_reporting}{Flag indicating whether to output number of reported severe and fatal cases}
}
\description{
Function to calculate annual yellow fever burden across multiple regions based on derived parameters
}
\details{
Function to take in parameter sets derived from MCMC fitting and use to calculate annual total and reported
case and death numbers for multiple regions to compare with external data
}
