% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{truncate_mcmc_data}
\alias{truncate_mcmc_data}
\title{truncate_mcmc_data}
\usage{
truncate_mcmc_data(input_frame = list(), rows = c(1), plot_graph = TRUE)
}
\arguments{
\item{input_frame}{Data frame of output data}

\item{rows}{Vector of line numbers to extract from input_frame}

\item{plot_graph}{TRUE/FALSE flag indicating whether to plot results graph
'}
}
\description{
Take MCMC results generated by get_mcmc_data and extract selected lines; optionally plot likelihood
graph and/or output data set
}
\details{
This function takes in a data frame output by the get_mcmc_data() function and outputs a data frame
consisting only of the selected lines; it can be used to cut off all pre-burnin points in the Markov chain and/or
to obtain a selection of points along part of the chain. A graph of the posterior likelihood values in the
truncated data frame can optionally be plotted.
}
